# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "客服系统api `\n客服后端：`",
        "title": "客服系统API文档`",
        "contact": {},
        "version": "1.0`"
    },
    "paths": {
...
        "/manager/auth/login": {
            "post": {
                "description": "原密码 admin 新密码 123456`",
                "tags": [
                    "客服系统总后台"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新密码",
                        "name": "new_password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/service/common/api/oss_config": {
            "post": {
                "tags": [
                    "客服系统",
                    "用户端"
                ],
                "summary": "获取上传参数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/service/common/count": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    }
}